+++ CHANGELOG.md
@@ -23,7 +23,7 @@
 
 ### Fixes
 
-- None
+- Fixed issue with `GraphQLResult` reprs - [#374](https://github.com/PrefectHQ/prefect/pull/374)
 
 ### Breaking Changes